Rename `diff_edit.rs` to `merge_tools.rs`
This is important especially since we're about to rename the `touchup` command into `diffedit`.
