+++ src/lib.rs
@@ -18,10 +18,10 @@ pub mod cleanup_guard;
 pub mod cli_util;
 pub mod commands;
 pub mod config;
-pub mod diff_edit;
 pub mod diff_util;
 pub mod formatter;
 pub mod graphlog;
+pub mod merge_tools;
 mod progress;
 pub mod template_parser;
 pub mod templater;
